requirements.txt
spacy==2.3.2
tqdm==4.45.0
torchvision==0.6.1
torch==1.5.1
pandas==1.0.3
Pillow==7.2.0
